requirements.txt
# Core LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-mistralai>=0.1.0

# LCEL / Runnable support (comes via langchain-core but keeping explicit)
langchainhub>=0.1.15

# OpenAI
openai>=1.30.0

# Tavily Search Tool
tavily-python>=0.3.3

# Web scraping
beautifulsoup4>=4.12.3
lxml>=5.2.1
requests>=2.32.0
httpx>=0.27.0

# Async support (important for agents + tools)
aiohttp>=3.9.5

# Data handling
pydantic>=2.7.0

# Environment variables
python-dotenv>=1.0.1

# Logging & utilities
tenacity>=8.2.3

# Optional but useful (debugging + dev)
rich>=13.7.1
tqdm>=4.66.4

# JSON handling improvements
orjson>=3.10.0

# Optional: FastAPI (if you plan to expose API)
fastapi>=0.111.0
uvicorn>=0.30.0

# Optional: Jupyter for testing
ipython>=8.24.0